Remove potentially existent aliases
Before creating symbolic links to aliases, remove targets first.
Notice that we cannot use `ln -sf` for some reason; the `-f` flag
seems to be ignored on a few tests I have run.

This is an issue when we are building on an image that already
has `$INSTALL_DEST/$ALIAS`, because `ln -s ...` is ineffective,
and the alias still points to a potentially nonexistent directory.